# core/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.ExportRequest:
    properties:
      format:
        description: Format is "bundle" (default) or "markdown".
        type: string
      operationId:
        type: string
      rootId:
        description: RootId scopes the export to one page and its descendants.
        type: string
    required:
    - operationId
    type: object
  controller.JobResponse:
    properties:
      artifactName:
        description: ArtifactName is the filename offered on download.
        type: string
      artifactSize:
        type: integer
      createAt:
        type: string
      error:
        description: Error is the request-level failure message when Status is failed.
        type: string
      expiresAt:
        type: string
      finishedAt:
        type: string
      format:
        $ref: '#/definitions/models.WikiTransferFormat'
      id:
        type: string
      jobId:
        type: string
      kind:
        $ref: '#/definitions/models.WikiTransferKind'
      operationId:
        type: string
      progress:
        $ref: '#/definitions/models.WikiTransferProgress'
      report:
        items:
          type: integer
        type: array
      request:
        $ref: '#/definitions/models.WikiTransferRequest'
      requestedById:
        type: string
      startedAt:
        type: string
      status:
        $ref: '#/definitions/models.WikiTransferStatus'
      updateAt:
        type: string
    type: object
...
  models.WikiTransferFormat:
    enum:
    - bundle
    - markdown
    type: string
    x-enum-varnames:
    - WikiTransferBundle
    - WikiTransferMarkdown
  models.WikiTransferKind:
    enum:
    - export
    - import
    type: string
    x-enum-varnames:
    - WikiTransferExport
    - WikiTransferImport
  models.WikiTransferProgress:
    properties:
      done:
        type: integer
      total:
        type: integer
    type: object
  models.WikiTransferRequest:
    properties:
      holdingPen:
        description: |-
          HoldingPen lands an import in import/<timestamp>/ instead of
          TargetParentID.
        type: boolean
      includeCredentials:
        description: IncludeCredentials embeds credential payloads in an export.
        type: boolean
      rootDocumentId:
        description: RootDocumentID scopes an export to a subtree; nil exports the
          tree.
        type: string
      targetParentId:
        description: TargetParentID places an import under a page; nil means the root.
        type: string
      uploadFilename:
        description: UploadFilename is the name of the uploaded archive, for display.
        type: string
    type: object
  models.WikiTransferStatus:
    enum:
    - queued
    - running
    - done
    - failed
    type: string
    x-enum-varnames:
    - WikiTransferQueued
    - WikiTransferRunning
    - WikiTransferDone
    - WikiTransferFailed
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: Command-and-control core backend API.
  title: VibeC2 Core API
  version: "1.0"
paths:
...
  /wiki/transfer/exports:
    post:
      consumes:
      - application/json
      description: Queues an export of the operation's wiki (or one subtree) as a
        native bundle or a markdown zip. Poll the returned job and download the archive
        when it is done. Viewer+ on the operation; credentials are embedded only when
        the caller is operator+.
      parameters:
      - description: Export request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.ExportRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a wiki export job
      tags:
      - Wiki
  /wiki/transfer/imports:
    post:
      consumes:
      - multipart/form-data
      description: Accepts a native bundle (.vibewiki.zip) or a markdown zip. The
        format is detected from the archive. Pages land under `targetParentId`, at
        the root when it is absent, or in import/<timestamp>/ when `holdingPen=true`.
        Operator+ on the operation.
      parameters:
      - description: Target operation ID (UUID)
        in: query
        name: operationId
        required: true
        type: string
      - description: Page to import under (UUID). Omit for the root.
        in: query
        name: targetParentId
        type: string
      - description: Land in import/<timestamp>/ instead of targetParentId
        in: query
        name: holdingPen
        type: boolean
      - description: Archive
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload an archive and start a wiki import job
      tags:
      - Wiki
  /wiki/transfer/jobs:
    get:
      parameters:
      - description: Operation ID (UUID)
        in: query
        name: operationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.JobResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List an operation's wiki transfer jobs
      tags:
      - Wiki
  /wiki/transfer/jobs/{id}:
    get:
      parameters:
      - description: Job ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JobResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fetch one wiki transfer job
      tags:
      - Wiki
  /wiki/transfer/jobs/{id}/download:
    get:
      parameters:
      - description: Job ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download a finished export
      tags:
      - Wiki
securityDefinitions:
  BearerAuth:
    description: Bearer JWT token (e.g. "Bearer eyJhb...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"